refactor(ui): flatten component folder structure
