+++ src/features/canvas-editing/components/menu-bar/CanvasResetViewButton.tsx
