+++ src/features/canvas-editing/components/menu-bar/CanvasZoomInButton.tsx
