+++ src/features/canvas-editing/components/menu-bar/CanvasZoomOutButton.tsx
