+++ src/features/canvas-editing/components/tool-bar/AddObstacleButton.tsx
