+++ src/features/canvas-editing/components/tool-bar/AddZoneButton.tsx
