+++ src/features/canvas-editing/components/tool-bar/DeleteButton.tsx
