+++ src/features/canvas-editing/components/tool-bar/EditButton.tsx
