+++ src/features/coverage-planning/components/controls-panel/algo-section/AlgoSelect.tsx
