+++ src/features/coverage-planning/components/controls-panel/algo-section/PathWidthInput.tsx
