+++ src/features/coverage-planning/components/controls-panel/env-section/GlobalTypeSelect.tsx
