+++ src/features/coverage-planning/components/controls-panel/object-section/ObjectTypeSelect.tsx
