+++ src/features/coverage-planning/components/tool-bar/ClearCppButton.tsx
