+++ src/features/coverage-planning/components/tool-bar/ExecuteCppButton.tsx
