+++ src/layouts/workbench-window/application-header/menu-bar/view-submenu/MenuBarViewSubmenu.tsx
@@ -1,7 +1,7 @@
-import CanvasResetViewButton from "@/features/canvas-editing/components/menu-bar/CanvasResetViewButton/CanvasResetViewButton";
-import CanvasZoomInButton from "@/features/canvas-editing/components/menu-bar/CanvasZoomInButton/CanvasZoomInButton";
-import CanvasZoomOutButton from "@/features/canvas-editing/components/menu-bar/CanvasZoomOutButton/CanvasZoomOutButton";
-import CanvasToggleGridButton from "@/features/canvas-editing/components/menu-bar/CanvasToggleGridButton/CanvasToggleGridButton";
+import CanvasResetViewButton from "@/features/canvas-editing/components/menu-bar/CanvasResetViewButton";
+import CanvasZoomInButton from "@/features/canvas-editing/components/menu-bar/CanvasZoomInButton";
+import CanvasZoomOutButton from "@/features/canvas-editing/components/menu-bar/CanvasZoomOutButton";
+import CanvasToggleGridButton from "@/features/canvas-editing/components/menu-bar/CanvasToggleGridButton";
 import MenuSeparator from "@/components/menu-bar/MenuBarSeparator";
 import UiInspectorButton from "@/features/ui-manager/components/menu-bar/UiInspectorButton/UiInspectorButton";
 import UiControlsButton from "@/features/ui-manager/components/menu-bar/UiControlsButton/UiControlsButton";
