+++ src/layouts/workbench-window/controls-panel/env-section/EnvSection.tsx
@@ -1,6 +1,6 @@
 import ControlsPanelSection from "@/components/controls-panel/ControlsPanelSection";
-import GlobalTypeSelection from "@/features/coverage-planning/components/controls-panel/EnvSection/GlobalTypeSelect/GlobalTypeSelect";
-import FormatSelection from "@/features/coverage-planning/components/controls-panel/EnvSection/FormatSelect/FormatSelect";
+import GlobalTypeSelection from "@/features/coverage-planning/components/controls-panel/env-section/GlobalTypeSelect";
+import FormatSelection from "@/features/coverage-planning/components/controls-panel/env-section/FormatSelect";
 
 export default function EnvSection() {
   return (
