+++ src/layouts/workbench-window/controls-panel/object-section/ObjectSection.tsx
@@ -1,5 +1,5 @@
 import ControlsPanelSection from "@/components/controls-panel/ControlsPanelSection";
-import ObjectTypeSelect from "@/features/coverage-planning/components/controls-panel/ObjectSection/ObjectTypeSelect/ObjectTypeSelect";
+import ObjectTypeSelect from "@/features/coverage-planning/components/controls-panel/object-section/ObjectTypeSelect";
 
 export default function ObjectSection() {
   return (
